Eliminate dynamic test creation for parallel sqlite test
